# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OM-Paie API",
        "description": "API de paiement mobile et transfert d'argent - OM-Paie avec MongoDB",
        "contact": {
            "email": "support@om-paie.sn"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "https://tinkin-transfer.onrender.com/api",
            "description": "Serveur de production sur Render"
        },
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur de développement"
        }
    ],
    "paths": {
...
        "/comptes/{numeroCompte}": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Obtenir les informations d'un compte",
                "description": "Récupère les informations d'un compte spécifique (solde, etc.)",
                "operationId": "getCompteInfo",
                "parameters": [
                    {
                        "name": "numeroCompte",
                        "in": "path",
                        "description": "Numéro du compte (format: Principal{timestamp} ou Secondaire{timestamp})",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Principal1731580000"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Informations du compte récupérées avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Informations du compte récupérées avec succès"
                                        },
                                        "data": {
                                            "properties": {
                                                "numero_compte": {
                                                    "type": "string",
                                                    "example": "a05ab654-1a7e-47ca-a9f0-3af7926e5e82"
                                                },
                                                "solde": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 1947500
                                                },
                                                "devise": {
                                                    "type": "string",
                                                    "example": "XOF"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé - compte n'appartient pas à l'utilisateur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Compte non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/comptes": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Lister tous les comptes de l'utilisateur",
                "description": "Récupère la liste de tous les comptes (principal et secondaires) de l'utilisateur authentifié",
                "operationId": "getUserComptes",
                "responses": {
                    "200": {
                        "description": "Comptes récupérés avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Comptes récupérés avec succès"
                                        },
                                        "comptes": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "numero_compte": {
                                                        "type": "string",
                                                        "example": "a05ab654-1a7e-47ca-a9f0-3af7926e5e82"
                                                    },
                                                    "solde": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 1947500
                                                    },
                                                    "devise": {
                                                        "type": "string",
                                                        "example": "XOF"
                                                    },
                                                    "type": {
                                                        "type": "string",
                                                        "example": "principal"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorResponse": {
                "title": "Réponse d'erreur",
                "description": "Réponse standard en cas d'erreur",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Message d'erreur"
                    },
                    "error": {
                        "type": "string",
                        "example": "error_code",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "OAuth2 avec Laravel Passport",
                "flows": {
                    "password": {
                        "tokenUrl": "/oauth/token",
                        "scopes": {}
                    }
                }
            },
            "bearerAuth": {
                "type": "http",
                "description": "Bearer Token (JWT) obtenu après connexion",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}